Typescriptify wrld.js (main file only)
