--- .npmignore
@@ -1,4 +1,0 @@
-# Files not be included in the npm package
-/coverage
-test.sh
-issue_template.md
