+++ src/types/buildings.ts
