+++ src/types/indoorMapEntities.ts
