+++ src/types/indoorMapFloorOutlines.ts
