+++ src/types/module.ts
@@ -1,0 +1,5 @@
+export type Module = {
+  ctx: unknown;
+  locateFile: (url: string) => string;
+  onExit: (code: number) => void;
+};
