+++ src/types/window.ts
@@ -1,0 +1,9 @@
+import L from "leaflet";
+import { Module } from "./module";
+
+declare global {
+  interface Window {
+    createWrldModule: (module: Module) => Module;
+    L: typeof L;
+  }
+}
