+++ tsconfig.json
@@ -1,12 +1,21 @@
 {
   "compilerOptions": {
-    "outDir": "dist",
-    "module": "esnext",
+    "module": "ESNext",
     "lib": ["dom", "esnext"],
     "moduleResolution": "node",
     "jsx": "react",
-    "sourceMap": true,
-    "declaration": true,
+    "target": "ESNext",
+    // "sourceMap": true,
+    // "declaration": true,
+    // "noEmit": false,
+    "types": ["jest", "node"],
+    // interop configuration
+    "skipLibCheck": true,
+    "allowSyntheticDefaultImports": true,
+    "allowJs": true,
+    "stripInternal": true,
+    // type checking
+    "strict": true,
     "esModuleInterop": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
@@ -15,18 +24,10 @@
     "suppressImplicitAnyIndexErrors": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "allowSyntheticDefaultImports": true,
-    "target": "es5",
-    "allowJs": true,
-    "skipLibCheck": true,
-    "strict": true,
     "forceConsistentCasingInFileNames": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
     "noFallthroughCasesInSwitch": true,
-    "stripInternal": true,
-    "types": ["jest", "node"]
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist", "example"]
